+++ CardGame/CardGame/CardDeck.swift
@@ -9,20 +9,21 @@
 import Foundation
 
 class CardDeck {
+    private var cardDeck : [Card]
+    init() { cardDeck = [] }
     
+    // 카드 덱 초기화
+    func reset() {
+    }
+    
+    // 카드 덱의 현재 카드 개수를 반환
     func count() -> Int {
-        return 10
+        return cardDeck.count
     }
     
     func shuffle() {
-        
     }
     
     func removeOne() {
-        
-    }
-    
-    func reset() {
-        
     }
 }
