+++ CardGame/CardGame/Number.swift
@@ -1,0 +1,40 @@
+//
+//  Number.swift
+//  CardGame
+//
+//  Created by 윤동민 on 04/12/2018.
+//  Copyright © 2018 JK. All rights reserved.
+//
+
+import Foundation
+
+enum Number : Int, CustomStringConvertible {
+    case one = 1
+    case two = 2
+    case three = 3
+    case four = 4
+    case five = 5
+    case six = 6
+    case seven = 7
+    case eight = 8
+    case nine = 9
+    case ten = 10
+    case eleven = 11
+    case twelve = 12
+    case thirteen = 13
+    
+    var description: String {
+        switch self {
+        case .one:
+            return "A"
+        case .eleven:
+            return "J"
+        case .twelve:
+            return "Q"
+        case .thirteen:
+            return "K"
+        default:
+            return "\(self.rawValue)"
+        }
+    }
+}
